requirements.txt
streamlit>=1.30.0
requests>=2.31.0
openai>=1.50.0
python-dotenv>=1.0.0
